requirements.txt
anyio==3.6.1
APScheduler==3.9.1
astropy==5.1
astroquery==0.4.6
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
fastapi==0.85.0
fonttools==4.37.4
h11==0.12.0
html5lib==1.1
httpcore==0.14.7
httpx==0.22.0
idna==3.4
jaraco.classes==3.2.3
jplephem==2.18
keyring==23.9.3
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.6.0
more-itertools==8.14.0
numba==0.56.2
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
plotly==5.10.0
poliastro==0.17.0
pydantic==1.10.2
pyerfa==2.0.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.0a0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pyvo==1.4
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
schedule==1.1.0
scipy==1.9.1
sgp4==2.21
six==1.16.0
skyfield==1.45
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.20.4
tenacity==8.1.0
TLE-tools==0.2.3
tornado==6.1
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.18.3
webencodings==0.5.1
